--- a/lib/order/index.js
+++ b/lib/order/index.js
@@ -34,6 +34,10 @@ function order(g) {
   var layering = initOrder(g);
   assignOrder(g, layering);
 
+  if (crossCount(g, layering) == 0) {
+    return;
+  }
+
   var bestCC = Number.POSITIVE_INFINITY,
     best;